% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_presence_of_required_columns.R
\name{check_presence_of_required_columns}
\alias{check_presence_of_required_columns}
\title{Verify if the required columns are present in the dataset}
\usage{
check_presence_of_required_columns(df)
}
\arguments{
\item{df}{Data frame. A dataset with squirrels data.}
}
\value{
a boolean
}
\description{
long, lat and primary_fur_color are required
}
\examples{

check_presence_of_required_columns(df = squirrels::data_demo_squirrels)
}
