% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_squirrel_data_integrity.R
\name{check_squirrel_data_integrity}
\alias{check_squirrel_data_integrity}
\title{Function to check squirrel data integrity}
\usage{
check_squirrel_data_integrity(data_df)
}
\arguments{
\item{data_df}{Squirrel dataset}
}
\value{
String returns a message with the data integrity information
}
\description{
Function to check squirrel data integrity
}
\examples{
data_path <- system.file("nyc_squirrels_act_sample.csv", package="squirrels")
data_act_squirrels <- readr::read_csv(data_path)
check_squirrel_data_integrity(data_act_squirrels)
}
