% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_df.R
\docType{data}
\name{data_df}
\alias{data_df}
\title{data_df}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{ age }{  character }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  integer }
}
}
\source{
Source
}
\usage{
data_df
}
\description{
Some dummy dataset for teaching purpose.
}
\keyword{datasets}
