% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df_squirrels.R
\docType{data}
\name{df_squirrels}
\alias{df_squirrels}
\title{df_squirrels}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{ age }{  character }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  numeric }
}
}
\source{
Source
}
\usage{
df_squirrels
}
\description{
Description.
}
\keyword{datasets}
