% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_squirrels_data_based_on_integrity.R
\name{filter_squirrels_data_based_on_integrity}
\alias{filter_squirrels_data_based_on_integrity}
\title{Vérifie l'intégrité d'un data frame}
\usage{
filter_squirrels_data_based_on_integrity(data)
}
\arguments{
\item{data}{un data frame}
}
\value{
un message : df_valid_data : données “valides” au niveau des couleurs et des coordonnées, df_invalid_data : données “invalides” au niveau des couleurs et des coordonnées, df_missing_data : données pour lesquelles la validité n’a pu être vérifiée
}
\description{
Vérifie l'intégrité d'un data frame
}
\examples{
filter_squirrels_data_based_on_integrity(squirrels::data_demo_squirrels)
}
