% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{Title Calcul du nombres de lignes d'un dataframe}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
un message
}
\description{
Title Calcul du nombres de lignes d'un dataframe
}
\examples{
get_df_nrows(df = iris)
get_df_nrows(df = mtcars)
}
