% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_color.R
\name{get_message_fur_color}
\alias{get_message_fur_color}
\title{Send a message with the squirrel color of interest}
\usage{
get_message_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{A character string corresponding to primary fur color.}
}
\value{
Nothing. It is used for its side effect of sending a message.
}
\description{
Send a message with the squirrel color of interest
}
\examples{
get_message_fur_color(primary_fur_color = "Black")
}
