% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_colour.R
\name{get_message_fur_colour}
\alias{get_message_fur_colour}
\title{my_fun Title}
\usage{
get_message_fur_colour(primary_fur_color)
}
\value{
1
}
\description{
my_fun Title
}
\examples{
get_message_fur_colour("Cinnamon")
}
