% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_paul.R
\name{get_message_paul}
\alias{get_message_paul}
\title{my_fun Title}
\usage{
get_message_paul(date = NULL)
}
\arguments{
\item{date}{une date au format character}
}
\value{
un message
}
\description{
my_fun Title
}
\examples{
# get_message_paul()
# get_message_paul("26/01/2023")

}
