% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_welcome_message.R
\name{get_welcome_message}
\alias{get_welcome_message}
\title{Envoie un message de bienvenue avec la date choisie par l'utilisateur}
\usage{
get_welcome_message(date)
}
\arguments{
\item{date}{character qui prend la forme \%d/\%m/\%y}
}
\value{
glue character
}
\description{
Envoie un message de bienvenue avec la date choisie par l'utilisateur
}
\examples{
date_ok <- "01/12/2023"
get_welcome_message(date_ok)
}
