% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_welcome_message_benoit.R
\name{get_welcome_message_benoit}
\alias{get_welcome_message_benoit}
\title{get_welcome_message_benoit}
\usage{
get_welcome_message_benoit(date)
}
\arguments{
\item{date}{A \code{string} character providing the date to be included in the welcome message.}
}
\value{
A message in the console welcoming the user with the provided \code{date} parameter.
}
\description{
Benoit's function to send a welcome message to the user.
}
\examples{
get_welcome_message_benoit("02/01/1997")

}
