% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_welcome_message_pierre.R
\name{get_welcome_message_pierre}
\alias{get_welcome_message_pierre}
\title{my_fun get_welcome_message_pierre}
\usage{
get_welcome_message_pierre(date)
}
\arguments{
\item{date}{une date au format date}
}
\value{
Un message ou une erreur
}
\description{
Renvoie un message avec une date entrée par l'utilisateur
}
\examples{
get_welcome_message_pierre(as.Date("23/01/2023", "\%d/\%m/\%Y"))
}
