% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_welcome_message_raul.R
\name{get_welcome_message_raul}
\alias{get_welcome_message_raul}
\title{get_welcome_message_raul Title}
\usage{
get_welcome_message_raul(date = stringr::str_replace_all(Sys.Date(), "-", "/"))
}
\arguments{
\item{date}{a character string that looks like a date}
}
\value{
A message
}
\description{
get_welcome_message_raul Title
}
\examples{
get_welcome_message_raul()
get_welcome_message_raul(date="09/02/1982")

}
