% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nyc_squirrels_act.R
\docType{data}
\name{nyc_squirrels_act}
\alias{nyc_squirrels_act}
\title{nyc_squirrels_act}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{ age }{  character - l age des ecureuils }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  numeric }
}
}
\source{
Source
}
\usage{
nyc_squirrels_act
}
\description{
Description.
}
\keyword{datasets}
