% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nyc_squirrels_act_sample_15.R
\docType{data}
\name{nyc_squirrels_act_sample_15}
\alias{nyc_squirrels_act_sample_15}
\title{nyc_squirrels_act_sample_15}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{ age }{  character }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  integer }
}
}
\source{
R
}
\usage{
nyc_squirrels_act_sample_15
}
\description{
The first 15 lines of nyc_squirrels_act_sample.
}
\keyword{datasets}
