% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nyc_squirrels_dummy.R
\docType{data}
\name{nyc_squirrels_dummy}
\alias{nyc_squirrels_dummy}
\title{nyc_squirrels_dummy}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{ age }{  character }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  numeric }
}
}
\source{
Source
}
\usage{
nyc_squirrels_dummy
}
\description{
A data set representing squirrels survey in New York.
}
\keyword{datasets}
