% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nyc_squirrels_reduced.R
\docType{data}
\name{nyc_squirrels_reduced}
\alias{nyc_squirrels_reduced}
\title{nyc_squirrels_reduced}
\format{
A data frame with 25 rows and 3 variables:
\describe{
\item{ long }{  numeric }
\item{ lat }{  numeric }
\item{ primary_fur_color }{  character }
}
}
\source{
Thinkr
}
\usage{
nyc_squirrels_reduced
}
\description{
Echantillon de 25 observations d'écureuils new-yorkais.
}
\keyword{datasets}
