% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nyc_squirrels_sample.R
\docType{data}
\name{nyc_squirrels_sample}
\alias{nyc_squirrels_sample}
\title{nyc_squirrels_sample}
\format{
A data frame with 25 rows and 3 variables:
\describe{
\item{ long }{  numeric }
\item{ lat }{  numeric }
\item{ primary_fur_color }{  character }
}
}
\source{
Données fournies par Thinkr via Bakacode (dossier "data").
}
\usage{
nyc_squirrels_sample
}
\description{
Description.
}
\keyword{datasets}
