% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_day_squirrels.R
\name{read_data_day_squirrels}
\alias{read_data_day_squirrels}
\title{Read the data of a given day of observation}
\usage{
read_data_day_squirrels(path)
}
\arguments{
\item{path}{Character. Path to a folder containing files of the form "nyc_squirrels_dd.xlsx"}
}
\value{
A named list. Each element corresponds to the data of a given day. The element is NULL if the file is corrupted.
}
\description{
Read the data of a given day of observation
}
\examples{
library(purrr)
lst_fic <- list.files(
path = system.file(package = "squirrels"),
pattern = "nyc_squirrels_[0-9][0-9].xlsx",
full.names = TRUE)
lst_dir <- map_chr(.x = lst_fic, .f = dirname) |> unique()
squirrels <- read_data_day_squirrels(lst_dir)
}
