% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the analysis}
\usage{
run_analysis(df, folder_in_which_to_save_csv_files)
}
\arguments{
\item{df}{a data frame}

\item{folder_in_which_to_save_csv_files}{character. folder in which to save the csv files}
}
\value{
a message
}
\description{
Run the analysis
}
\examples{
data("data_demo_squirrels")
output_dir <- tempfile()
dir.create(output_dir)
run_analysis(df = data_demo_squirrels, folder_in_which_to_save_csv_files = output_dir)
unlink(output_dir, recursive = TRUE)
}
