% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Sauvegarde d'un df en csv}
\usage{
save_as_csv(df, path)
}
\arguments{
\item{df}{df a exporter}

\item{path}{chemin de sortie}
}
\value{
le chemin de sortie
}
\description{
Sauvegarde d'un df en csv
}
\examples{
mon_dossier_temp <- tempfile(pattern ="savecsv")
dir.create(mon_dossier_temp)
#dans la console : rstudio::filesPanelNaviguate(mon_dossier_temp)
save_as_csv(df=iris, path=file.path(mon_dossier_temp, "output.csv"))
#Ménage :
unlink(mon_dossier_temp, recursive = TRUE)
}
