% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrels-package.R
\docType{package}
\name{squirrels-package}
\alias{squirrels}
\alias{squirrels-package}
\title{squirrels: Calcul Du Nombre De Lignes}
\description{
Cette fonction sert à calculer le nombre de lignes d'un dataframe.
}
\author{
\strong{Maintainer}: Delphine \email{dhocqquette@atih.sante.fr} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Balbla \email{blabla@atih.sante.fr} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
