% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Checks if the coordinates are in Central Park}
\usage{
check_coordinates(latitude, longitude)
}
\arguments{
\item{latitude}{numeric. A vector with the latitude.}

\item{longitude}{numeric. A vector with the longitude.}
}
\value{
a vector of logicals
}
\description{
Checks if the coordinates are in Central Park
}
\examples{
both_valid_latitude <- c(40.766, 40.799)
both_valid_longitude <- c(-73.96, -73.95)
check_coordinates(
latitude = both_valid_latitude,
longitude = both_valid_longitude
)
}
