% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Fonction qui vérifie si les coordonnées sont correctes}
\usage{
check_coordinates()
}
\description{
Fonction qui vérifie si les coordonnées sont correctes
}
\examples{
check_coordinates()
}
