% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plus_in_primary_fur_color.R
\name{check_plus_in_primary_fur_color}
\alias{check_plus_in_primary_fur_color}
\title{Fonction qui vérifie que la couleur entrée est bien un chaine de caractère}
\usage{
check_plus_in_primary_fur_color(primary_fur_color)
}
\arguments{
\item{vec}{un vecteur de couleur}
}
\value{
une erreur si le vecteur n’est pas de type “character”
}
\description{
Fonction qui vérifie que la couleur entrée est bien un chaine de caractère
}
\examples{
mymtcars <- head(iris)
check_plus_in_primary_fur_color(as.character(mymtcars$Species))
}
