% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_presence_of_required_columns}
\alias{check_presence_of_required_columns}
\title{check the presence of required columns}
\usage{
check_presence_of_required_columns(dataframe)
}
\description{
Vérifie que le jeu de données continet bien les colonnes requises pour l'étude des écureuils de Central Park
}
\examples{
data_is_ok <- data_demo_squirrels
check_presence_of_required_columns(data_is_ok)
data_not_ok <- iris
check_presence_of_required_columns(data_not_ok)
}
