% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_presence_of_required_columns.R
\name{check_presence_of_required_columns}
\alias{check_presence_of_required_columns}
\title{Vérifie que toutes les colonnes existent}
\usage{
check_presence_of_required_columns(data)
}
\arguments{
\item{un}{data frame}
}
\value{
un booléen TRUE si tout est OK, et un message d’erreur dans le cas contraire
}
\description{
Vérifie que toutes les colonnes existent
}
\examples{
check_presence_of_required_columns(squirrels::data_demo_squirrels)
check_presence_of_required_columns(mtcars)

}
