% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primary_color_is_ok.R
\name{check_primary_color_is_ok}
\alias{check_primary_color_is_ok}
\title{Check the values of primary fur color}
\usage{
check_primary_color_is_ok(string)
}
\arguments{
\item{string}{Character. A vector with the primary fur color.}
}
\value{
Boolean. TRUE if all colors are correct.
}
\description{
Check the values of primary fur color
}
\examples{
check_primary_color_is_ok("Cinnamon")
}
