% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_squirrel_data_integrity.R
\name{check_squirrel_data_integrity}
\alias{check_squirrel_data_integrity}
\title{Check the integrity of the column of primary fur colors}
\usage{
check_squirrel_data_integrity(df_squirrels)
}
\arguments{
\item{df_squirrels}{Data frame. A dataset with squirrels data.}
}
\value{
Nothing. Use for side effect. Displays a message if everything is ok
}
\description{
Check the integrity of the column of primary fur colors
}
\examples{
my_data_squirrels <- readr::read_csv(
  system.file("nyc_squirrels_sample.csv", package = "squirrels")
  )
check_squirrel_data_integrity(my_data_squirrels)
}
