% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_demo_squirrels.R
\docType{data}
\name{data_demo_squirrels}
\alias{data_demo_squirrels}
\title{data_demo_squirrels}
\format{
A data frame with 25 rows and 3 variables:
\describe{
\item{ long }{  numeric }
\item{ lat }{  numeric }
\item{ primary_fur_color }{  character }
}
}
\source{
Source
}
\usage{
data_demo_squirrels
}
\description{
Description.
}
\keyword{datasets}
