% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_squirrels_data_based_on_integrity.R
\name{filter_squirrels_data_based_on_integrity}
\alias{filter_squirrels_data_based_on_integrity}
\title{Filter the data based on the quality of the primary fur color and the coordinates}
\usage{
filter_squirrels_data_based_on_integrity(data)
}
\arguments{
\item{data}{Data frame. A dataset with squirrels data.}
}
\value{
a list with valid, invalid, missing data
}
\description{
Filter the data based on the quality of the primary fur color and the coordinates
}
\examples{
filter_squirrels_data_based_on_integrity(data = data_demo_squirrels)


my_data <- readr::read_csv(system.file("nyc_squirrels_sample.csv", package = "squirrels"))
filter_squirrels_data_based_on_integrity(data = my_data)

}
