% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_squirrels_data_based_on_integrity.R
\name{filter_squirrels_data_based_on_integrity}
\alias{filter_squirrels_data_based_on_integrity}
\title{Filter the data based on the quality of the primary fur color and the coordinates}
\usage{
filter_squirrels_data_based_on_integrity(df_squirrels)
}
\arguments{
\item{df_squirrels}{Data frame. A dataset with squirrels data.}
}
\value{
A list with 3 elements (df_valid_data, df_invalid_data, df_missing_data) based on their integrity (color and coordinates)
}
\description{
Filter the data based on the quality of the primary fur color and the coordinates
}
