% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{To get a message with the number of rows in a data frame of interest with \code{get_df_nrows()}.}
\usage{
get_df_nrows(data)
}
\arguments{
\item{data}{dataframe The dataset to be used}
}
\value{
numeric (the number of rows)  and a message
}
\description{
To get a message with the number of rows in a data frame of interest with \code{get_df_nrows()}.
}
\examples{
get_df_nrows(data = iris)
}
