% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{get_df_nrows}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{DataFrame to work with}
}
\value{
The number of rows in df
}
\description{
Get a message with the number of rows in a dataframe
}
\examples{
get_df_nrows(df = mtcars)
}
