% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{Get a message with number of rows of the dataset you will be working with}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{data.frame. The dataset to be used}
}
\value{
Used for side effect. Outputs a message in the console
}
\description{
Get a message with number of rows of the dataset you will be working with
}
\examples{
get_df_nrows(df = iris)
}
