% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{Fonction qui affiche le nombre de lignes}
\usage{
get_df_nrows(data)
}
\arguments{
\item{data}{le data frame qui est utilisé}
}
\value{
un numerique (nombre de lignes) et un message
}
\description{
Fonction qui affiche le nombre de lignes
}
\examples{
get_df_nrows(data=iris)
get_df_nrows(data=mtcars)
}
