% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{get_df_nrows}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
nrow Number of rows
}
\description{
Compute number of rows
}
\examples{
get_df_nrows(df = mtcars)
}
