% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_color.R
\name{get_message_fur_color}
\alias{get_message_fur_color}
\title{my_fun Title}
\usage{
get_message_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{primary color of interest}
}
\value{
Used for side effects. output a message in the console
}
\description{
my_fun Title
}
