% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{export_results Title}
\usage{
save_as_csv(df, mypath, ...)
}
\arguments{
\item{df}{A \code{data_frame}}

\item{input_path}{}
}
\value{
full_path A \code{string} indicating the full path of the created file.
}
\description{
export_results Title
}
\examples{
example_path <- file.path(tempfile(pattern="example_csv"), "example1.csv")
dirname(example_path)
normalizePath(example_path)
save_as_csv(iris, example_path)
extra_args <- c(col.names = FALSE, na = "NA_modified")


save_as_csv(iris, example_path, col.names = FALSE, na = "NA_modified")
}
