% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Sauvegarde d'un df en csv}
\usage{
save_as_csv(df, path)
}
\arguments{
\item{df}{df a exporter}

\item{path}{chemin de sortie}
}
\value{
le chemin de sortie
}
\description{
Sauvegarde d'un df en csv
}
\examples{
chemin_tmp <- tempfile(pattern = "output")
dir.create(chemin_tmp)
save_as_csv(data_demo_squirrels, path =file.path(chemin_tmp, "iris.csv"))
unlink(chemin_tmp, recursive = TRUE)


}
