% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save as csv}
\usage{
save_as_csv(df, mypath)
}
\arguments{
\item{df}{a data.frame}

\item{mypath}{le chemin ou enregistrer mon df}
}
\value{
Le chemin d enregistrement
}
\description{
Save as csv
}
\examples{

path <- "mondossier"

mon_dossier_temp <- tempfile(pattern = "saveascsv")
dir.create(mon_dossier_temp)
mypath <- file.path(mon_dossier_temp, "output.csv")

save_as_csv(df = iris, mypath = mypath)
unlink(mypath)
}
