% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{A function to save csv}
\usage{
save_as_csv(dataset, path_name)
}
\arguments{
\item{dataset}{A data frame}

\item{path_name}{A file path}
}
\value{
Doesn't eturn anything writes a file to disk
}
\description{
A function to save csv
}
\examples{

library(dplyr)
# Create a temporary directory
my_temp_folder <- tempfile(pattern = "savecsv")
dir.create(my_temp_folder)

data(iris)
# ok: the .csv file is saved in the current project
iris \%>\% save_as_csv(file.path(my_temp_folder, "output.csv"))

# ok: the .csv file is saved and the file opens
iris \%>\% save_as_csv(file.path(my_temp_folder, "output.csv")) \%>\% browseURL() 
}
