% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{export  file}
\usage{
save_as_csv(df, mypath, ...)
}
\arguments{
\item{df}{}

\item{mypath}{le chemin où enregistrer mon df}

\item{...}{}
}
\value{
le chemin d'enregistrement
}
\description{
save as CSV
}
\examples{

path <- "mondossier"

mon_dossier_temp <- tempfile(pattern = "saveascsv")
dir.create(mon_dossier_temp)
mypath <- file.path(mon_dossier_temp, "output.csv")

save_as_csv(df = iris, mypath = mypath)
unlink(mypath)
}
