% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{export_results Title}
\usage{
save_as_csv(df, mypath)
}
\arguments{
\item{df}{dataframe}

\item{path}{le chemin où enregistrer mon df}
}
\value{
le chemin d'enregistrement
}
\description{
export_results Title
}
\examples{
export_results()
}
