% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Sauvegarde d'un df en csv}
\usage{
save_as_csv(df, path)
}
\arguments{
\item{df}{df à exporter}

\item{path}{chemin de sortie}
}
\value{
le chemin de sortie
}
\description{
Sauvegarde d'un df en csv
}
\examples{
dossier_temp <- tempfile(pattern = "saveascsv")
dir.create(dossier_temp)
save_as_csv(df = data_demo_squirrels, path = file.path(dossier_temp, "monfichier.csv"))
unlink(dossier_temp, recursive = TRUE)

}
