% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save an output as a csv file}
\usage{
save_as_csv(df, mypath, ...)
}
\arguments{
\item{df}{Data frame. The table to be saved as a csv file.}

\item{mypath}{Character. Path where the csv file shloud be saved.}

\item{...}{other parameters}
}
\value{
The path to the created csv file.
}
\description{
Save an output as a csv file
}
\examples{

mypath <- "mondossier"

mon_dossier_temp <- tempfile(pattern = "saveascsv")
dir.create(mon_dossier_temp)
mypath <- file.path(mon_dossier_temp, "output.csv")

save_as_csv(df = iris, mypath = mypath)
unlink(mypath)
}
