% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Une fonction qui prend en paramètres un jeu de données et une chaîne de caractères qui correspond au chemin du fichier qui doit être créé  permet d’enregistrer le jeu de données en csv avec la fonction write.csv2() du package {utils} au chemin spécifié renvoie en sortie le chemin complet du csv créé}
\usage{
save_as_csv(data, string)
}
\arguments{
\item{data}{un datafram}

\item{string}{un chemin}
}
\description{
Description
}
\examples{
mon_dossier_temp_aleatoire <- tempfile(pattern = "toto")
dir.create(mon_dossier_temp_aleatoire) # crée le dossier temporaire
file.create(file.path(mon_dossier_temp_aleatoire, save_as_csv(iris, "blabla")))
}
