% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Exportations de données}
\usage{
save_as_csv(df, chemin)
}
\value{
Le chemin ou se trouve le fichier créé
}
\description{
Cette fonction permet d'exporter les données en format csv
}
\examples{
data("data_demo_squirrels")
save_as_csv(data_demo_squirrels,"data.csv")
}
