% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrels-package.R
\docType{package}
\name{squirrels-package}
\alias{squirrels}
\alias{squirrels-package}
\title{squirrels: Study The Squirrels Of Central Park}
\description{
Provide everything need to study the squirrels' population in Central Park, NYC.
}
\author{
\strong{Maintainer}: Magali Dauvin \email{magali.dauvin@sciencespo.fr}

}
\keyword{internal}
