% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrels-package.R
\docType{package}
\name{squirrels-package}
\alias{squirrels}
\alias{squirrels-package}
\title{squirrels: Analyze Squirrels Data}
\description{
This package provides tools to analyse squirrels of NYC.
}
\author{
\strong{Maintainer}: Murielle Delmotte \email{murielle@thinkr.fr}

}
\keyword{internal}
